изменения

+++ src/main/java/com/businessplanner/models/Task.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class Task {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(nullable = false)
     private String title;
@@ -37,6 +37,8 @@ public class Task {
         joinColumns = @JoinColumn(name = "task_id"),
         inverseJoinColumns = @JoinColumn(name = "tag_id")
     )
+
+    
     private List<Tag> tags;
 
     public List<TaskTag> getTaskTags() {
